requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.8
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
narwhals==1.30.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
plotly==6.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
wheel==0.45.1
whitenoise==6.9.0
